machine-learning/models/research/object_detection/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_tflow_select=2.2.0=eigen
absl-py=0.12.0=py36haa95532_0
aiohttp=3.7.4=py36h2bbff1b_1
argon2-cffi=20.1.0=pypi_0
astor=0.8.1=py36haa95532_0
async-generator=1.10=pypi_0
async-timeout=3.0.1=py36haa95532_0
attrs=20.3.0=pyhd3eb1b0_0
backcall=0.2.0=pypi_0
blas=1.0=mkl
bleach=3.3.0=pypi_0
blinker=1.4=py36haa95532_0
brotlipy=0.7.0=py36h2bbff1b_1003
ca-certificates=2021.4.13=haa95532_1
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2020.12.5=py36haa95532_0
cffi=1.14.5=py36hcd4344a_0
chardet=3.0.4=py36haa95532_1003
click=7.1.2=pyhd3eb1b0_0
colorama=0.4.4=pypi_0
contextlib2=0.6.0.post1=pypi_0
coverage=5.5=py36h2bbff1b_2
cryptography=3.4.7=py36h71e12ea_0
cycler=0.10.0=pypi_0
cython=0.29.23=py36hd77b12b_0
decorator=5.0.7=pypi_0
defusedxml=0.7.1=pypi_0
entrypoints=0.3=pypi_0
freetype=2.10.4=hd328e21_0
gast=0.2.2=py36_0
google-auth=1.30.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.4=pyhd3eb1b0_0
google-pasta=0.2.0=py_0
grpcio=1.36.1=py36hc60d5dd_1
h5py=2.10.0=py36h5e291fa_0
hdf5=1.10.4=h7ebc959_0
icc_rt=2019.0.0=h0cc432a_1
idna=2.10=pyhd3eb1b0_0
idna_ssl=1.1.0=py36haa95532_0
importlib-metadata=3.10.0=py36haa95532_0
intel-openmp=2021.2.0=haa95532_616
ipykernel=5.5.3=pypi_0
ipython=7.16.1=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.6.3=pypi_0
jedi=0.18.0=pypi_0
jinja2=2.11.3=pypi_0
jpeg=9b=hb83a4c4_2
jsonschema=3.2.0=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=6.1.12=pypi_0
jupyter-console=6.4.0=pypi_0
jupyter-core=4.7.1=pypi_0
jupyterlab-pygments=0.1.2=pypi_0
jupyterlab-widgets=1.0.0=pypi_0
keras-applications=1.0.8=py_1
keras-preprocessing=1.1.2=pyhd3eb1b0_0
kiwisolver=1.3.1=pypi_0
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.14.0=h23ce68f_0
libtiff=4.2.0=hd0e1b90_0
lvis=0.5.3=pypi_0
lxml=4.6.3=pypi_0
lz4-c=1.9.3=h2bbff1b_0
markdown=3.3.4=py36haa95532_0
markupsafe=1.1.1=pypi_0
matplotlib=3.3.4=pypi_0
mistune=0.8.4=pypi_0
mkl=2020.2=256
mkl-service=2.3.0=py36h196d8e1_0
mkl_fft=1.3.0=py36h46781fe_0
mkl_random=1.1.1=py36h47e9c7a_0
multidict=5.1.0=py36h2bbff1b_2
nbclient=0.5.3=pypi_0
nbconvert=6.0.7=pypi_0
nbformat=5.1.3=pypi_0
nest-asyncio=1.5.1=pypi_0
notebook=6.3.0=pypi_0
numpy=1.19.2=py36hadc3359_0
numpy-base=1.19.2=py36ha3acd2a_0
oauthlib=3.1.0=py_0
object-detection=0.1=pypi_0
olefile=0.46=py36_0
opencv-python=4.5.1.48=pypi_0
openssl=1.1.1k=h2bbff1b_0
opt_einsum=3.1.0=py_0
packaging=20.9=pypi_0
paho-mqtt=1.5.1=pypi_0
pandas=1.1.5=pypi_0
pandocfilters=1.4.3=pypi_0
parso=0.8.2=pypi_0
pickleshare=0.7.5=pypi_0
pillow=8.2.0=py36h4fa10fc_0
pip=21.0.1=py36haa95532_0
prometheus-client=0.10.1=pypi_0
prompt-toolkit=3.0.18=pypi_0
protobuf=3.14.0=py36hd77b12b_1
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=py_0
pycocotools=2.0.2=pypi_0
pycparser=2.20=py_2
pygments=2.8.1=pypi_0
pyjwt=1.7.1=py36_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=pypi_0
pyreadline=2.1=py36_1
pyrsistent=0.17.3=pypi_0
pysocks=1.7.1=py36haa95532_0
python=3.6.13=h3758d61_0
python-dateutil=2.8.1=pypi_0
pytz=2021.1=pypi_0
pywin32=300=pypi_0
pywinpty=1.0.1=pypi_0
pyzmq=22.0.3=pypi_0
qtconsole=5.0.3=pypi_0
qtpy=1.9.0=pypi_0
requests=2.25.1=pyhd3eb1b0_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
scipy=1.1.0=pypi_0
send2trash=1.5.0=pypi_0
setuptools=52.0.0=py36haa95532_0
six=1.15.0=py36haa95532_0
sqlite=3.35.4=h2bbff1b_0
tensorboard=1.15.0=pypi_0
tensorboard-plugin-wit=1.6.0=py_0
tensorflow=1.15.0=eigen_py36h932cce6_0
tensorflow-base=1.15.0=eigen_py36h07d2309_0
tensorflow-estimator=1.15.1=pyh2649769_0
termcolor=1.1.0=py36haa95532_1
terminado=0.9.4=pypi_0
testpath=0.4.4=pypi_0
tf-slim=1.1.0=pypi_0
tk=8.6.10=he774522_0
tornado=6.1=pypi_0
traitlets=4.3.3=pypi_0
typing-extensions=3.7.4.3=hd3eb1b0_0
typing_extensions=3.7.4.3=pyh06a4308_0
urllib3=1.26.4=pyhd3eb1b0_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
werkzeug=0.16.1=py_0
wheel=0.36.2=pyhd3eb1b0_0
widgetsnbextension=3.5.1=pypi_0
win_inet_pton=1.1.0=py36haa95532_0
wincertstore=0.2=py36h7fe50ca_0
wrapt=1.12.1=py36he774522_1
xz=5.2.5=h62dcd97_0
yarl=1.6.3=py36h2bbff1b_0
zipp=3.4.1=pyhd3eb1b0_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.5=h04227a9_0
